Colorings, cycles, double triangles
Extracted data for ways of coloring the graphs using different number of colors, number of cycles of different length and number of double triangles. Based on initial checks with decision trees and correlation matrix, the first two pieces of info do not look promising. Still have to check the double triangles.