Adjacency matrix solved
Both vector and scalars properties incorporated and ran the model. The config_adj.yaml, config_adj_2.yaml determine the training to be done either on just the adjacency matrix or on all the features. Initially there is a lot of overfitting and the validation error does not reduce. Adjusting the hyperparameters (simplifying the model by reducing numer of hidden channels, layers, increasing dropout and weight decay) reduced this problem. The adjacency info is probably too much.