Binary classification adjustment
Adjusted the code to be suitable for binary classification. The output layer is now one dimensional and BCEWithLogitsLoss is used as loss function. Moreover the ROC AUC is printed by wandb for each epoch. Having this metric allows us to compare results with other methods. ROC AUC is threshold independent. Running few examples we get similar results with before.