Extra training data
Incorporated the extra data derived from the cusp rule into the model training. The option extra_training in the config specifies whether the extra graphs from den_graph_data_7to*.csv will be considered in the training or not. Tried 7->8 but the results were worse with the extra graphs. Still investigating why.