Introduced validation set
We want to optimize the hyperparameters of our model. Doing this by just checking on the training loss does not seem correct. Moreover, doing that on 80% of the training set still does not seem correct since training and test sets have different number of nodes and the model will be biased against lower number of nodes. Instead of testing on test_loop and train on 6,..,test_loop-1 I train on 6,..,test_loop-2+70% of test_loop-1, validate on 30% of test_loop-1 and finally test on test_loop.