Introduced ansatz reduction metric
The metric corresponds to the safest reduction of the ansatz that the model provides. This is computed by first evaluating the model on the test set, picking the smallest value given by the model to a true 1 and setting it as a threshold. All true 1s will be inside it but also many false 1s. The percentage of the ansatz reduction is printed.