Create rerun_best.py
Function that takes as input a sweep project id, a base config file and a tag, and runs the corresponding config files for 100 epochs. Will be used to automatize the second step of hyperparameter choice. Each run creates a new config file to be used for training that is also saved.